web/pyproject.toml
[tool.poetry]
name = "book-shop"
version = "0.1.0"
description = ""
authors = ["Sergey <sergeymaksheev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "book_shop"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"
python-dotenv = {extras = ["cli"], version = "^0.21.0"}
djangorestframework = "^3.14.0"
djoser = "^2.1.0"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.5"

[build-system]
requires = ["poetry"]
build-backend = "poetry.core.masonry.api"
